refactor: move arbitrary code exec manager into dedicated module
