+++ src/griptape_nodes/retained_mode/griptape_nodes.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
-import io
 import logging
 import re
-from contextlib import redirect_stdout
 from dataclasses import dataclass
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING, Any, ClassVar, TextIO, TypeVar
@@ -20,11 +18,6 @@ from griptape_nodes.retained_mode.events.app_events import (
     GetEngineVersionRequest,
     GetEngineVersionResultFailure,
     GetEngineVersionResultSuccess,
-)
-from griptape_nodes.retained_mode.events.arbitrary_python_events import (
-    RunArbitraryPythonStringRequest,
-    RunArbitraryPythonStringResultFailure,
-    RunArbitraryPythonStringResultSuccess,
 )
 from griptape_nodes.retained_mode.events.base_events import (
     AppPayload,
@@ -43,6 +36,7 @@ from griptape_nodes.retained_mode.events.parameter_events import (
     AddParameterToNodeRequest,
     AlterParameterDetailsRequest,
 )
+from griptape_nodes.retained_mode.managers.arbitrary_code_exec_manager import ArbitraryCodeExecManager
 from griptape_nodes.retained_mode.managers.config_manager import ConfigManager
 from griptape_nodes.retained_mode.managers.context_manager import ContextManager
 from griptape_nodes.retained_mode.managers.event_manager import EventManager
@@ -517,27 +511,6 @@ def manage_alter_details(parameter: Parameter, base_node_obj: BaseNode) -> dict:
     else:
         return vars(parameter)
     return diff
-
-
-class ArbitraryCodeExecManager:
-    def __init__(self, event_manager: EventManager) -> None:
-        event_manager.assign_manager_to_request_type(
-            RunArbitraryPythonStringRequest, self.on_run_arbitrary_python_string_request
-        )
-
-    def on_run_arbitrary_python_string_request(self, request: RunArbitraryPythonStringRequest) -> ResultPayload:
-        try:
-            string_buffer = io.StringIO()
-            with redirect_stdout(string_buffer):
-                python_output = exec(request.python_string)  # noqa: S102
-
-            captured_output = string_buffer.getvalue()
-            result = RunArbitraryPythonStringResultSuccess(python_output=captured_output)
-        except Exception as e:
-            python_output = f"ERROR: {e}"
-            result = RunArbitraryPythonStringResultFailure(python_output=python_output)
-
-        return result
 
 
 def __getattr__(name) -> logging.Logger:
